+++ resources/views/sections/PRCPST01001/section.blade.php
@@ -1,6 +1,6 @@
 <section id="{{ $section['id'] }}" class="switchable {{ checkAdjustable() }} {{ checkManageable() }} {{ parseSecAttr('.adjustable', $section['classes']) }}" {{ parseEditSecAttr($page['meta']['mode'],$section) }} {{ parseSecAttr('.adjustable', $section['attributes']) }}>
     <div class="container">
-        <div class="row justify-content-between">
+        <div class="row justify-content-between {{ parseSecAttr('.row', $section['classes']) }}">
             <div class="col-md-6 adaptable__text">
 	            @if(checkContent($section['PRCPST02001_HEAD'], 'title'))
 			    	<h2>{{ parseContent($section['PRCPST02001_HEAD'],'title') }}</h2>